Simulator: Basic implementation
Menu: Two Paged, controlled by tab. One menu is used to
select the algorithm and the second one is where the simulation
occurs

Simulator window contains Header, description, and the array. All
the inputs are provided in the javascript file, and the information
from this file is taken and used in the simulation

Signed-off by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>